# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FLOW API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/publishing/snapshots/{report_snapshot_id}/publish": {
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Publish Report",
        "operationId": "publish_report_api_v1_publishing_snapshots__report_snapshot_id__publish_post",
        "parameters": [
          {
            "name": "report_snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Snapshot Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PublishRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishingErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishingErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publishing/snapshots/{report_snapshot_id}/attempts": {
      "get": {
        "tags": [
          "publishing"
        ],
        "summary": "Publication Attempts",
        "operationId": "publication_attempts_api_v1_publishing_snapshots__report_snapshot_id__attempts_get",
        "parameters": [
          {
            "name": "report_snapshot_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Snapshot Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicationAttemptsResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishingErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publishing/snapshots": {
      "post": {
        "tags": [
          "publishing"
        ],
        "summary": "Freeze Report Snapshot Route",
        "description": "从已发布指标快照与已批准 findings 冻结报告快照（append-only 身份）。",
        "operationId": "freeze_report_snapshot_route_api_v1_publishing_snapshots_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportSnapshotFreezeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSnapshotCreatedResponse"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishingErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "publishing"
        ],
        "summary": "List Report Snapshots",
        "operationId": "list_report_snapshots_api_v1_publishing_snapshots_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSnapshotListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publishing/freeze-candidates": {
      "get": {
        "tags": [
          "publishing"
        ],
        "summary": "List Freeze Candidates",
        "description": "已发布指标快照及已批准 Finding 数，供冻结表单选择。",
        "operationId": "list_freeze_candidates_api_v1_publishing_freeze_candidates_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FreezeCandidateListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/publishing/attempts/{attempt_id}/download": {
      "get": {
        "tags": [
          "publishing"
        ],
        "summary": "Download Publication Attempt",
        "description": "下载成功产物的持久化字节：服务端命名 + sha 校验 + no-store/nosniff。",
        "operationId": "download_publication_attempt_api_v1_publishing_attempts__attempt_id__download_get",
        "parameters": [
          {
            "name": "attempt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Attempt Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishingErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublishingErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorDetail": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "additionalProperties": true,
            "type": "object",
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail"
      },
...
      "FreezeCandidateLine": {
        "properties": {
          "metric_snapshot_id": {
            "type": "string",
            "title": "Metric Snapshot Id"
          },
          "batch_id": {
            "type": "string",
            "title": "Batch Id"
          },
          "period_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Period Label"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "approved_findings": {
            "type": "integer",
            "title": "Approved Findings"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "metric_snapshot_id",
          "batch_id",
          "period_label",
          "version",
          "approved_findings"
        ],
        "title": "FreezeCandidateLine",
        "description": "可冻结的已发布指标快照：报告中心冻结表单的选择项。"
      },
      "FreezeCandidateListResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/FreezeCandidateLine"
            },
            "type": "array",
            "title": "Candidates"
          }
        },
        "type": "object",
        "required": [
          "candidates"
        ],
        "title": "FreezeCandidateListResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PublicationAttemptLine": {
        "properties": {
          "attempt_id": {
            "type": "string",
            "title": "Attempt Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "stored_object_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored Object Id"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "content_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Type"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "download_available": {
            "type": "boolean",
            "title": "Download Available",
            "default": false
          },
          "stored_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stored Sha256"
          }
        },
        "type": "object",
        "required": [
          "attempt_id",
          "sequence",
          "format",
          "status",
          "stored_object_id",
          "error_message"
        ],
        "title": "PublicationAttemptLine"
      },
      "PublicationAttemptsResponse": {
        "properties": {
          "report_snapshot_id": {
            "type": "string",
            "title": "Report Snapshot Id"
          },
          "attempts": {
            "items": {
              "$ref": "#/components/schemas/PublicationAttemptLine"
            },
            "type": "array",
            "title": "Attempts"
          }
        },
        "type": "object",
        "required": [
          "report_snapshot_id",
          "attempts"
        ],
        "title": "PublicationAttemptsResponse"
      },
      "PublishRequest": {
        "properties": {
          "formats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Formats"
          },
          "actor": {
            "type": "string",
            "minLength": 1,
            "title": "Actor"
          }
        },
        "type": "object",
        "required": [
          "actor"
        ],
        "title": "PublishRequest"
      },
      "PublishResponse": {
        "properties": {
          "report_snapshot_id": {
            "type": "string",
            "title": "Report Snapshot Id"
          },
          "outcomes": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Outcomes"
          }
        },
        "type": "object",
        "required": [
          "report_snapshot_id",
          "outcomes"
        ],
        "title": "PublishResponse"
      },
      "PublishingErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "PublishingErrorResponse"
      },
...
      "ReportSnapshotCreatedResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "metric_snapshot_id": {
            "type": "string",
            "title": "Metric Snapshot Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "metric_snapshot_id",
          "version",
          "title"
        ],
        "title": "ReportSnapshotCreatedResponse"
      },
      "ReportSnapshotFreezeRequest": {
        "properties": {
          "metric_snapshot_id": {
            "type": "string",
            "minLength": 1,
            "title": "Metric Snapshot Id"
          }
        },
        "type": "object",
        "required": [
          "metric_snapshot_id"
        ],
        "title": "ReportSnapshotFreezeRequest"
      },
      "ReportSnapshotLine": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "metric_snapshot_id": {
            "type": "string",
            "title": "Metric Snapshot Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "metric_snapshot_id",
          "version",
          "title"
        ],
        "title": "ReportSnapshotLine"
      },
      "ReportSnapshotListResponse": {
        "properties": {
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/ReportSnapshotLine"
            },
            "type": "array",
            "title": "Snapshots"
          }
        },
        "type": "object",
        "required": [
          "snapshots"
        ],
        "title": "ReportSnapshotListResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}